[Refactor/#8] header 및 footer 파일 분리
